+++ src/main/java/it/gov/pagopa/register/configuration/EprelValidationConfig.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.register.config;
+package it.gov.pagopa.register.configuration;
 
 import it.gov.pagopa.register.utils.EprelValidationRule;
 import lombok.Data;
