+++ src/main/java/it/gov/pagopa/register/configuration/ProductFileValidationConfig.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.register.config;
+package it.gov.pagopa.register.configuration;
 
 
 import it.gov.pagopa.register.utils.ColumnValidationRule;
