+++ src/main/java/it/gov/pagopa/register/service/operation/ProductFileService.java
@@ -2,7 +2,7 @@ package it.gov.pagopa.register.service.operation;
 
 import it.gov.pagopa.register.connector.storage.FileStorageClient;
 import it.gov.pagopa.register.constants.AssetRegisterConstants;
-import it.gov.pagopa.register.constants.enums.UploadCsvStatus;
+import it.gov.pagopa.register.enums.UploadCsvStatus;
 import it.gov.pagopa.register.dto.operation.*;
 import it.gov.pagopa.register.exception.operation.ReportNotFoundException;
 import it.gov.pagopa.register.mapper.operation.ProductFileMapper;
@@ -31,7 +31,7 @@ import java.util.List;
 import java.util.Objects;
 
 import static it.gov.pagopa.register.constants.AssetRegisterConstants.*;
-import static it.gov.pagopa.register.constants.enums.UploadCsvStatus.*;
+import static it.gov.pagopa.register.enums.UploadCsvStatus.*;
 
 @Slf4j
 @Service
