+++ src/test/java/it/gov/pagopa/register/connector/eprel/EprelConnectorTest.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.register.connector.eprel;
 
-import it.gov.pagopa.register.config.RestTemplateConfig;
+import it.gov.pagopa.register.configuration.RestTemplateConfig;
 import it.gov.pagopa.register.utils.EprelProduct;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
